Binomial tree pricing for a Stock Option
implementation of a Binomial tree pricing for an American put option (assuming non-dividend paying Stock). Step1 is to generate the entire tree with future stock prices based on risk free rate, volatility, spot price, maturity and # of steps in tree. The tree is a vector<vector<double>>. Step2 is to do a backward induction on the tree to calculate the American price (based on node price and strike) and European price (based on option price at the leaves of the node and probability of up/down movement) at each node. Max of American/European price is the price at each node and price at node 0 is the total option price.